vehicle-signal-server-spec/0150-get-error-invalid-path.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: get method tests</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<p> Get method test: error case 01: invalid path</p>
<p> Get Signal.Drivetrain.Transmission.xxxxxx</p>
<p> (* Run VSSS server before starting test) </p>

<h3>
<div id="log"></div>
</h3>

<script>
var t = async_test("Get method error: invalid data path");

var vehicle  = new WebSocket(VISS_URL);

vehicle.onopen = function() {
  // request a signal
  var reqId = "reqid-12345";
  var path = "Signal.Drivetrain.Transmission.xxxxxx";

  // process messages from the server
  vehicle.onmessage = t.step_func(
    function(event) {
      var msg = JSON.parse(event.data);

      if( isGetErrorResponse(reqId, msg) ) {
        addLogMessage("getErrorResponse received. err_number = " + msg.error.number);
        addLogSuccess("getErrorResponse returned as expected");
        assert_true(true, "OK: error as expected: Err: "+msg.error.number+":"+msg.error.reason +", path = " +path);
      } else {
        addLogFailure("getErrorResponse was expected");
      }
      t.step_timeout(function() {
        vehicle.close();
        t.done();
      }, TIME_FINISH_WAIT);
    }
  );
  // in case server doesn't return response
  t.step_timeout(function() {
    addLogFailure("Fishish by timeout");
    t.step_timeout(function() {
      assert_true(false);
      vehicle.close();
      t.done();
    }, TIME_FINISH_WAIT);
  }, 3000);
  addLogMessage("Send get request: path="+path);
  vehicle.send('{"action":"get","path":"'+ path +'" ,"requestId":"'+reqId+'"}');
}
</script>
